Refactoring _PlotWidgetTest into a utils module
- The class is now renamed PlotWidgetTestCase
